Converts a slew of AvmCore methods to static. Performance increase is modest in JIT mode on most benchmarks, but math-centric benchmarks (eg primes, crypto) show 15-25% improvement in JIT mode. Also added some minor but worthwhile optimizations to the bodies of AvmCore::number, AvmCore::istype, Toplevel::add2. (r=lhansen)

+++ nanojit/LIR.cpp
@@ -1730,7 +1730,7 @@ namespace nanojit
 			VMPI_strcat(buf, cname8.c_str());
 		}
 		else if (ref->isconstq()) {
-            VMPI_sprintf(buf, "#0x%llxLL", (int64_t) ref->constvalq());
+            VMPI_sprintf(buf, "#0x%llxLL", (long long unsigned int) ref->constvalq());
 		}
 		else if (ref->isconst()) {
 			formatImm(ref->constval(), buf);
